+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Quonfig Go SDK are documented here.
 
-## Unreleased
+## 1.2.0 - 2026-07-08
 
 ### Added